remove validation
I don't know why exporter_exporter tries to reason about the metrics in
the first place. Prometheus is able to decide itself what it can
eat or not.

Also the verification library is pinned to version 0.2.0. The current
one is 0.31.1 <https://github.com/prometheus/common/tree/v0.31.1/expfmt>

We were trying to proxy to a jmx_exporter and it spit out the following
line:

And exporter_exporter reacted with an internal server error and
complained about not knowing the type "info".